Database update
There is now a database done through the built in sqlite. The longer term idea is to host the DB on a raspberry pi. This will suffice for testing purposes until I can get a pi and set it up. There is also now a password match error check during the registration